+++ src/nvhtmlib/novx_to_html.py
@@ -1,4 +1,4 @@
-"""Provide a class for parsing novx section content, converting it to XHTML.
+"""Provide a class for parsing novx section content, converting it to HTML.
 
 Copyright (c) 2025 Peter Triesberger
 For further information see https://github.com/peter88213/novx_html
@@ -7,21 +7,23 @@ License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 from xml import sax
 
 
-class NovxToXhtml(sax.ContentHandler):
-    """A parser to convert novx markup to XHTML."""
-    NON_XHTML = ('comment', 'note')
+class NovxToHtml(sax.ContentHandler):
+    """A parser to convert novx markup to HTML."""
+    NON_HTML = ('comment', 'note')
 
     def __init__(self):
         super().__init__()
         self.textList = None
         self._skip = None
+        self._firstParagraphInChapter = None
 
-    def feed(self, xmlString):
+    def feed(self, xmlString, firstInChapter):
         """Feed a string file to the parser.
         
         Positional arguments:
             filePath: str -- novx document path.        
         """
+        self._firstParagraphInChapter = firstInChapter
         xmlString = f'<novx>{xmlString}</novx>'
         self.textList = []
         if xmlString:
@@ -46,7 +48,7 @@ class NovxToXhtml(sax.ContentHandler):
         if name == 'novx':
             return
 
-        if name in self.NON_XHTML:
+        if name in self.NON_HTML:
             self._skip = False
             return
 
@@ -63,7 +65,7 @@ class NovxToXhtml(sax.ContentHandler):
         if name == 'novx':
             return
 
-        if name in self.NON_XHTML:
+        if name in self.NON_HTML:
             self._skip = True
             return
 
@@ -76,6 +78,9 @@ class NovxToXhtml(sax.ContentHandler):
             htmlAttributes.append(f'{attrKey}="{attrValue}"')
         if htmlAttributes:
             self.textList.append(f"<{name} {' '.join(htmlAttributes)}>")
+        elif name == 'p' and self._firstParagraphInChapter:
+            self.textList.append('<p class="ChpStart">')
+            self._firstParagraphInChapter = False
         else:
             self.textList.append(f'<{name}>')