+++ docs/changelog.md
@@ -5,6 +5,11 @@
 ## Changelog
 
 
+### Version 0.2.0
+
+- Assigning the "ChpStart" style to the first paragraph of a chapter.
+- Fixed sample templates.
+
 
 ### Version 0.1.0